Closes #61 and #62
- Override default subtraction to generate better firrtl.
- Add UInt typeclass.
- Some reorganization of existing typeclasses.
