+++ src/main/scala/dsptools/numbers/FixedPointImpl.scala
@@ -23,7 +23,10 @@ trait FixedPointRing extends Any with Ring[FixedPoint] with hasContext {
   }
   def one: FixedPoint = FixedPoint.fromBigInt(BigInt(1), binaryPoint = 0)
   def zero: FixedPoint = FixedPoint.fromBigInt(BigInt(0), binaryPoint = 0)
-  def negate(f: FixedPoint): FixedPoint = zero - f
+  def negate(f: FixedPoint): FixedPoint = -f
+  override def minus(f: FixedPoint, g: FixedPoint): FixedPoint = {
+    f - g
+  }
 }
 
 trait FixedPointImpl {
